+++ src/utils/findTableStartY.ts
@@ -1,5 +1,3 @@
-import getPixelRows from "./getPixelRows";
-
 export default function getTableStartY(imageData: ImageData) {
   const endOfFirstThird = Math.round(imageData.height / 3);
 
@@ -20,4 +18,16 @@ function findTableTop(rows: Uint8ClampedArray[]) {
   }
 
   return top;
+}
+
+function getPixelRows(imageData: ImageData) {
+  const lines = [];
+  const to = imageData.width * 4;
+
+  for (let index = 0; index * to < imageData.data.length; index++) {
+    const start = index * to;
+    lines.push(imageData.data.slice(start, start + to));
+  }
+
+  return lines;
 }
