+++ src/utils/setUpCanvas.ts
@@ -1,15 +1,24 @@
-export default function setUpCanvas(
-  onLoad: (
-    ctx: CanvasRenderingContext2D,
-    canvas: HTMLCanvasElement,
-    image: HTMLImageElement
-  ) => void
-) {
-  const canvas = document.querySelector("canvas")!;
+type LoadedPayload = {
+  ctx: CanvasRenderingContext2D;
+  canvas: HTMLCanvasElement;
+  image: HTMLImageElement;
+};
+
+export default function setUpCanvas() {
+  const canvas = document.createElement("canvas")!;
   const ctx = canvas.getContext("2d")!;
   const image = document.querySelector("img")!;
 
-  image.addEventListener("load", () => onLoad(ctx, canvas, image));
+  const promise = new Promise<LoadedPayload>((resolve, _reject) => {
+    image.addEventListener("load", () => {
+      canvas.width = image.width;
+      canvas.height = image.height;
 
-  image.src = "/images/TabelleStart_handled.png";
+      ctx.drawImage(image, 0, 0, image.width, image.height);
+      resolve({ ctx, canvas, image });
+    });
+    image.src = "/images/TabelleStart_handled.png";
+  });
+
+  return promise;
 }
